requirements.txt
Flask==2.0.1
Pillow==8.2.0
gunicon